// src/generated/payments/v1/payments.ts
export enum PaymentStatus {
  PAYMENT_STATUS_UNSPECIFIED = 0,
  PAYMENT_STATUS_AUTHORIZED = 1,
  PAYMENT_STATUS_CAPTURED = 2,
  PAYMENT_STATUS_REFUNDED = 3,
  PAYMENT_STATUS_FAILED = 4,
  UNRECOGNIZED = -1,
}